Adding tests for home page and share form page Renaming home page tests Adding proper vitest config to allow for automatic test cleanup Fixing jest DOM assertion errors
